backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
openai>=1.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
